Summary
It's a compiler satisfying expanded C0-grammars and the programming
language is C++. The project "Parser" and "ppp" in src directory might
be a little strange, but they're actually compilers.  "Parser" runs on
the Visual Studio 2013, while "ppp" runs on the Visual Studio 2010. I'm
sorry for their names might disturb you.